merge branch integration_2025-02-13_752315120898 into master

See merge request: !160

+++ src/Billing/Model/DeleteFinancialRelationRequest.php
@@ -41,7 +41,7 @@ class DeleteFinancialRelationRequest implements ModelInterface, ArrayAccess
     protected static $swaggerFormats = [
         'relation' => 'int32',
         'relation_id' => null,
-        'sub_account_id' => 'int32'
+        'sub_account_id' => 'int64'
     ];
 
     /**
